@@ -10,11 +10,13 @@ class BikeFactory extends Factory
      * Define the model's default state.
      *
      * @return array
+     *
+     * @todo Fixt the relation mapping on seeding the database. Else testing will not work as planed!
      */
     public function definition()
     {
         return [
-            'city_id' => $this->faker->numberBetween(1,6),
+            'city' => $this->faker->randomElement(["Stockholm", "Göteborg", "Malmö", "Karlskrona", "Umeå", "Västerås"]), // TODO FIX THIS!!!
             'status' => $this->faker->randomElement(["in_service", "in_repair", "broken", "lost", "out_of_service"]),
             'active' => $this->faker->boolean(),
             'longitude' => $this->faker->longitude(),
