@@ -4,6 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
+
 class CreateUsersTable extends Migration
 {
     /**
@@ -13,20 +14,21 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-//            $table->id()->autoIncrement()->unique();
-            $table->string('firstname');
-            $table->string('lastname');
-            $table->string('adress');
-            $table->string('postcode');
-            $table->foreignId('city_id');
-            $table->integer('phone')->unique();
-            $table->string('email')->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('password');
-            $table->string('payment_method');
-            $table->rememberToken();
-            $table->timestamps();
+        Schema::create('users', function (Blueprint $collection) {
+//            $collection->id()->autoIncrement()->unique();
+            $collection->string('firstname');
+            $collection->string('lastname');
+            $collection->string('adress');
+            $collection->string('postcode');
+            $collection->foreignId('city');
+            $collection->unique('phone');
+            $collection->unique('email');
+            $collection->timestamp('email_verified_at')->nullable();
+            $collection->string('password');
+            $collection->string('payment_method');
+            $collection->string('payment_status');
+            $collection->rememberToken();
+            $collection->timestamps();
         });
     }
 
