@@ -4,6 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
+
 class CreateTravelTable extends Migration
 {
     /**
@@ -13,22 +14,22 @@ class CreateTravelTable extends Migration
      */
     public function up()
     {
-        Schema::create('travel', function (Blueprint $table) {
-//            $table->id()->autoIncrement()->unique();
-            $table->foreignId('city_id');
-            $table->foreignId('user_id');
-            $table->foreignId('bike_id');
+        Schema::create('travel', function (Blueprint $collection) {
+            $collection->foreignId('city');
+            $collection->foreignId('user_id');
+            $collection->foreignId('bike_id');
 
-            $table->decimal('start_longitude', 10, 7)->nullable();
-            $table->decimal('start_latitude', 11, 8)->nullable();
-            $table->decimal('stop_longitude', 10, 7)->nullable();
-            $table->decimal('stop_latitude', 11, 8)->nullable();
+            $collection->decimal('start_longitude', 10, 7)->nullable();
+            $collection->decimal('start_latitude', 11, 8)->nullable();
+            $collection->decimal('stop_longitude', 10, 7)->nullable();
+            $collection->decimal('stop_latitude', 11, 8)->nullable();
 
-            $table->string('status')->default('active');
-            $table->integer('price')->default(10);
-            $table->timestamps();
+            $collection->string('status')->default('active');
+            $collection->integer('price')->default(10);
+            $collection->timestamps();
         });
     }
+
 
     /**
      * Reverse the migrations.
