@@ -4,6 +4,7 @@ namespace Database\Seeders;
 
 use Illuminate\Database\Seeder;
 
+
 class DatabaseSeeder extends Seeder
 {
     /**
@@ -13,43 +14,44 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        \App\Models\User::factory(25)->create();
+
+        \App\Models\User::factory(50)->create();
         \App\Models\Bike::factory(25)->create();
         \App\Models\Station::factory(10)->create();
 
 
         \App\Models\City::create([
-            'id' => 1,
+//            'id' => 1,
             'name' => 'Stockholm',
             'country' => 'Sverige'
         ]);
 
         \App\Models\City::create([
-            'id' => 2,
+//            'id' => 2,
             'name' => 'Göteborg',
             'country' => 'Sverige'
         ]);
 
         \App\Models\City::create([
-            'id' => 3,
+//            'id' => 3,
             'name' => 'Malmö',
             'country' => 'Sverige'
         ]);
 
         \App\Models\City::create([
-            'id' => 4,
+//            'id' => 4,
             'name' => 'Karlskrona',
             'country' => 'Sverige'
         ]);
 
         \App\Models\City::create([
-            'id' => 5,
+//            'id' => 5,
             'name' => 'Västerås',
             'country' => 'Sverige'
         ]);
 
         \App\Models\City::create([
-            'id' => 6,
+//            'id' => 6,
             'name' => 'Umeå',
             'country' => 'Sverige'
         ]);
